+++ i18n/src/main/res/values/ui.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="ui_title_foryou">M3U</string>
-    <string name="ui_title_favourite">Favourite</string>
+    <string name="ui_title_favourite">Favorites</string>
     <string name="ui_title_setting">Settings</string>
 
     <string name="ui_destination_foryou">For You</string>
-    <string name="ui_destination_favourite">Favourite</string>
+    <string name="ui_destination_favourite">Favorites</string>
     <string name="ui_destination_setting">Settings</string>
     <string name="ui_destination_playlist">Playlist</string>
